+++ lib/yw_clients/yw_clients/http/assets/__init__.py
@@ -1,0 +1,6 @@
+"""
+HTTP client & related for the :mod:`assets <youwol.backends.assets>` service.
+"""
+
+# relative
+from .assets import *