+++ lib/yw_clients/yw_clients/http/cdn/__init__.py
@@ -1,0 +1,6 @@
+"""
+HTTP client & related for the :mod:`cdn-backend <youwol.backends.cdn>` service.
+"""
+
+# relative
+from .cdn import *