+++ lib/yw_clients/yw_clients/http/explorer/__init__.py
@@ -1,0 +1,6 @@
+"""
+HTTP client & related for the :mod:`tree_db <youwol.backends.tree_db>` service.
+"""
+
+# relative
+from .explorer import *